app/requirements.txt
anyio==4.3.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
domain==0.1.4
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.89.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.34.0
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.15
PyMySQL==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.22.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.20.0
